+++ apps/web/app/entry.client.tsx
@@ -1,3 +1,4 @@
+import 'zod/compile';
 import { StrictMode, startTransition } from 'react';
 import { hydrateRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';